Move upload.js to data-capture and ES2015-ify it.
