+++ hourglass_site/static_source/js/data-capture/index.js
@@ -1,4 +1,5 @@
 
 // entrypoint for data-capture
 // this is where modules should be `require(...)`'d
-// example: require('../common/upload');
+
+require('./upload');
