+++ hourglass_site/templates/tests.html
@@ -18,6 +18,7 @@
   <body>
     <div id="qunit"></div>
     <div id="qunit-fixture"></div>
+    <script src="{% static 'hourglass_site/js/built/data-capture/index.min.js' %}"></script>
     <script src="{% static 'hourglass_site/tests/qunit-1.16.0.js' %}"></script>
     <script src="{% static 'hourglass_site/js/tests.js' %}"></script>
     <script src="{% static 'hourglass_site/js/upload_tests.js' %}"></script>
