+++ styleguide/templates/styleguide.html
@@ -179,4 +179,5 @@ if (upload.isDegraded) {
   {% endguide %}
 </div>
 <script src="{% static 'styleguide/vendor/prism.js' %}"></script>
+<script src="{% static 'hourglass_site/js/built/data-capture/index.min.js' %}"></script>
 {% endblock %}
